moved to gemm shared details to private header in src branch.
